Enable proper init self type by default (and reduce binary size)
BOOST_PYTHON_PY_SIGNATURES_PROPER_INIT_SELF_TYPE is not longer needed.
This was off by default because of the increase in module binary size.
The binary size was reduced by:

 * Using a CallPolicies signature metafunction instead of registering
an extra python_class<T> type.

 * Replacing most template instantiations of pytype_function with calls
to registry::lookup or registry::query.

 * Removing the static signature_element* initialisation. These are used
for docstrings and error reporting which is not performance critical,
even if the signatures need to be generated multiple times.